Function english_lint::lint

pub fn lint(input: &str) -> Vec<Hint>

Lint a string of english text

Given some english text, this function will try to give suggestions on how to improve the language.

Examples

extern crate english_lint;

let text: &str = "This chapter consists of relatively independent tutorials";
let suggestions: Vec<english_lint::Hint> = english_lint::lint(text);

assert_eq!(suggestions, vec![
    english_lint::Hint { group: Some("adverbs"),
                         value: "relatively".to_owned(),
                         line: Some(1), start: 25, end: 35 },
]);